backend-flask-server/requirements.txt
Flask
flask-cors
langchain
python-dotenv
langchain-openai
langchain-community